model GoogleMapsGeocodingDAO
